requirements.txt
boto3==1.26.94
botocore==1.29.94
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-config-file==0.6.0
configobj==5.0.8
dataclasses-json==0.5.4
decorator==5.1.1
graphql-core==3.2.3
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
montecarlodata==0.41.3
mypy-extensions==1.0.0
packaging==23.0
py==1.11.0
pycarlo==0.7.1
python-box==7.0.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.2
retry==0.9.2
s3transfer==0.6.0
sgqlc==14.1
six==1.16.0
stringcase==1.2.0
tabulate==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
